upload de imagens feito

+++ app/Http/Controllers/CategoryController.php
@@ -10,15 +10,22 @@ class CategoryController extends Controller
 {
     public function index(){
         $categories = Category::all();
-        return view('cat.index', ['categories' => $categories]);
+        if(Gate::allows('category')){
+            return view('cat.index', ['categories' => $categories]);
+        }else{
+            request()->session()->flash('alert-danger','Usuário sem permissão');
+            return redirect('/');
+        }
     }
     
     public function create(Category $category){
-        
+        if(Gate::allows('create-user')){
             return view('cat.create');
-        
-            
             return redirect('/cat');
+        }else{
+            request()->session()->flash('alert-danger','Usuário sem permissão');
+            return redirect('/');
+        }
     }
 
     public function store(Request $request, Category $category){
@@ -29,8 +36,13 @@ class CategoryController extends Controller
     }
 
     public function edit(Category $category){
-        return view('cat.edit', ['category' => $category]);
-        return redirect('/cat');
+        if(Gate::allows('create-user')){
+            return view('cat.edit', ['category' => $category]);
+            return redirect('/cat');
+        }else{
+            request()->session()->flash('alert-danger','Usuário sem permissão');
+            return redirect('/');
+        }
     }
 
     public function update(Request $request, Category $category){
@@ -40,7 +52,12 @@ class CategoryController extends Controller
     }
 
     public function destroy(Category $category){
-        $category->delete();
-        return redirect('/cat');
+        if(Gate::allows('create-user')){
+            $category->delete();
+            return redirect('/cat');
+        }else{
+            request()->session()->flash('alert-danger','Usuário sem permissão');
+            return redirect('/');
+        }
     }
 }
